+++ doc-examples/src/main/java/sting/doc/examples/interceptors/PaymentFragment.java
@@ -1,12 +1,10 @@
 package sting.doc.examples.interceptors;
 
 import sting.Fragment;
-import sting.Typed;
 
 @Fragment
 public interface PaymentFragment {
     @Timed
-    @Typed(PaymentGateway.class)
     default PaymentGateway providePaymentGateway() {
         return (accountId, amount) -> {};
     }