+++ processor/src/test/fixtures/input/com/example/fragment/types/RedundantTypedProviderMethodModel.java
@@ -1,0 +1,12 @@
+package com.example.fragment.types;
+
+import sting.Fragment;
+import sting.Typed;
+
+@Fragment
+public interface RedundantTypedProviderMethodModel {
+    @Typed(MyModel.class)
+    default MyModel provideMyModel() {
+        return null;
+    }
+}