+++ processor/src/test/fixtures/input/com/example/injectable/RedundantTypedAnnotationModel.java
@@ -1,0 +1,10 @@
+package com.example.injectable;
+
+import sting.Injectable;
+import sting.Typed;
+
+@Injectable
+@Typed(RedundantTypedAnnotationModel.class)
+public class RedundantTypedAnnotationModel {
+    RedundantTypedAnnotationModel() {}
+}