+++ processor/src/test/fixtures/input/com/example/injectable/SuppressedRedundantTypedAnnotationModel.java
@@ -1,0 +1,11 @@
+package com.example.injectable;
+
+import sting.Injectable;
+import sting.Typed;
+
+@SuppressWarnings("Sting:RedundantTypedAnnotation")
+@Injectable
+@Typed(SuppressedRedundantTypedAnnotationModel.class)
+public class SuppressedRedundantTypedAnnotationModel {
+    SuppressedRedundantTypedAnnotationModel() {}
+}